Push resolveCredentials and resolveRegion up
